style: improve code readability
