+++ src/router/index.js
@@ -1,7 +1,6 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
 
-import Firebase from 'firebase/app';
 import Routes from '@/router/routes';
 import store from '@/store';
 
@@ -12,18 +11,6 @@ const router = new VueRouter({
   routes: Routes,
 });
 
-router.beforeEach(async (to, from, next) => {
-  store.commit('setIsAuthLoading', true);
-  const isRequireAuth = to.matched.some((record) => record.meta.isRequireAuth);
-  const user = await Firebase.getCurrentUser();
-  if (isRequireAuth && !user) {
-    next('signin');
-  } else if (user && to.path === '/signin') {
-    next('/dashboard');
-  } else {
-    next();
-  }
-});
 // eslint-disable-next-line no-unused-vars
 router.afterEach((to, from) => {
   document.title = `${to.meta.title} - Kalgory Admin`;
